Privatize n2v and remove random walk iterator
